fix: address tlist mailbox review feedback
