+++ test/helpers/list.ts
@@ -1,0 +1,55 @@
+import { createTerminalApp } from "../../src/index.js";
+
+export function installRaf() {
+  const previousRaf = globalThis.requestAnimationFrame;
+  const previousCancel = globalThis.cancelAnimationFrame;
+  const callbacks = new Map<number, FrameRequestCallback>();
+  let rafId = 0;
+
+  globalThis.requestAnimationFrame = ((cb: FrameRequestCallback) => {
+    const id = ++rafId;
+    callbacks.set(id, cb);
+    return id;
+  }) as typeof requestAnimationFrame;
+  globalThis.cancelAnimationFrame = ((id: number) => {
+    callbacks.delete(id);
+  }) as typeof cancelAnimationFrame;
+
+  return {
+    callbacks,
+    runNext(time = 0): boolean {
+      const next = callbacks.entries().next().value;
+      if (!next) return false;
+      const [id, cb] = next;
+      callbacks.delete(id);
+      cb(time);
+      return true;
+    },
+    restore() {
+      globalThis.requestAnimationFrame = previousRaf;
+      globalThis.cancelAnimationFrame = previousCancel;
+    },
+  };
+}
+
+export function disableRaf() {
+  const g = globalThis as any;
+  const previousRaf = g.requestAnimationFrame;
+  const previousCancel = g.cancelAnimationFrame;
+  g.requestAnimationFrame = undefined;
+  g.cancelAnimationFrame = undefined;
+  return {
+    restore() {
+      g.requestAnimationFrame = previousRaf;
+      g.cancelAnimationFrame = previousCancel;
+    },
+  };
+}
+
+export function rowText(app: ReturnType<typeof createTerminalApp>, y: number): string {
+  return app.terminal
+    .getRow(y)
+    .map((cell) => cell.ch)
+    .join("")
+    .trimEnd();
+}
